requirements.txt
streamlit==1.39.0
pdfplumber==0.11.1
google-generativeai==0.8.3
python-dotenv==1.0.1
google-api-core==2.19.1
google-auth==2.35.0
protobuf==4.25.3
pillow==10.4.0
pypdfium2==4.30.0
pdfminer.six==20231228
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
packaging==24.1